Move all registering tests into a single file
